refactor `/about` to allow `onClick` handlers in `SocialLinks`
