+++ src/app/(main)/[mdxPageSlug]/about.mdx
@@ -14,10 +14,6 @@ In the past, I've worked extensively with Python and the Google Cloud Platform, 
 
 Outside of the world of programming, I am also a quizzer and was the lead at my college's [Quiz Club](https://www.facebook.com/qcnitw). Having played a lot of [CS](https://blog.counter-strike.net) in my childhood and college, I also love following the [CS:GO esports scene](https://www.hltv.org). I enjoy using mechanical keyboards and lurk [r/mk](https://www.reddit.com/r/MechanicalKeyboards) pretty much daily. I enjoy just about all genres of music and [podcasts](/podcasts) and am always looking for more ways to consume information from around the world. Additionally, I'm a [PADI certified](https://www.padi.com) scuba diver, equestrian and a pretty-decent cook.
 
-You can check me out on the following platforms:
-
-<ExternalLinksOverlay />
-
 If you're interested, you can take a look at my [resume](/resume.pdf) — and if you're intrigued still, thanks! You can reach out to me at sreetam [at] [`window.location.hostname`].
 
 Here's a list of pages that curate my website:
@@ -32,3 +28,5 @@ Here's a list of pages that curate my website:
 - [/faq](/faq) in case the question you have is already answered
 - [/credits](/credits) for the people who've helped and inspired me to make this website
 - [/rwc](/rwc) for solutions to the interview question of the week from the [rendezvous with @cassidoo](https://buttondown.email/cassidoo) newsletter (in TypeScript and Elixir so far!)
+
+You can check me out on the following platforms:
